# perf: optimize build process for parallel execution