# TL;DR

Build time improvement: `npm run vsix` (ie, `npm run build`) 

| before | after |
| ------ | ----- |
| 25.008s | 16.368s |

Please test, make sure this still works as you intend.  I think it does, and it works well for me. 

The commit is tiny, please review.

## Implementation


1. Package Script Parallelization:
   1. `build:webview`, `esbuild`, `check-types`, and `lint` are independent tasks
   2. Modified package script to run these in parallel using `npm-run-all`

2. Webview Build Optimization:
   - Split webview-ui build into separate tsc and vite-build tasks to run them in parallel

3. gzip stats:
   - Disabled vite's compressed size reporting which is slow (gzip), and (I think) it is only used for stats

## How to Test

approximately, this is my test script, edit as you see fit:

```sh
#!/bin/bash

set -e
set -x

npm run install:all

npm run vsix

# List resulting vsix files
pkg=$(ls -tr bin/roo-*.vsix | tail -n1)

echo "code --install-extension $pkg"
echo "code-insiders --install-extension $pkg"
```
<!-- ELLIPSIS_HIDDEN -->

----

> [!IMPORTANT]
> Optimized build process by parallelizing tasks and disabling gzip size reporting for faster execution.
> 
>   - **Build Process Optimization**:
>     - Parallelized `build:webview`, `build:esbuild`, `check-types`, and `lint` using `npm-run-all` in `package.json`.
>     - Split `webview-ui` build into `tsc` and `vite-build` tasks to run in parallel in `webview-ui/package.json`.
>   - **Performance Improvement**:
>     - Disabled `reportCompressedSize` in `webview-ui/vite.config.ts` to speed up build by skipping gzip size reporting.
> 
> <sup>This description was created by </sup>[<img alt="Ellipsis" src="https://img.shields.io/badge/Ellipsis-blue?color=175173">](https://www.ellipsis.dev?ref=RooVetGit%2FRoo-Code&utm_source=github&utm_medium=referral)<sup> for b04cd8f47a627f1be70e62e323aa2165c52e9a6a. It will automatically update as commits are pushed.</sup>

<!-- ELLIPSIS_HIDDEN -->